Refactor typings

+++ cards-backend/src/gateways/lobby.gateway.ts
@@ -1,3 +1,4 @@
+import { GameType } from '@models/game-type';
 import {
   ConnectedSocket,
   MessageBody,
@@ -7,11 +8,12 @@ import {
   WebSocketGateway,
   WebSocketServer,
 } from '@nestjs/websockets';
-import { GetLobbyPayload, JoinLobbyPayload } from '@shared';
 import { Server, Socket } from 'socket.io';
 import { ExtendedSocket } from 'src/interfaces/extended-socket';
 import { Lobby } from 'src/interfaces/lobby';
+import { OhHell } from 'src/interfaces/oh-hell';
 import { Player } from 'src/interfaces/player';
+import { GameService } from 'src/services/game/game.service';
 import { LobbyService } from 'src/services/lobby/lobby.service';
 
 @WebSocketGateway({
@@ -20,27 +22,34 @@ import { LobbyService } from 'src/services/lobby/lobby.service';
 export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect {
   @WebSocketServer() public server: Server;
 
-  constructor(private lobbyService: LobbyService) {}
+  constructor(private lobbyService: LobbyService, private gameService: GameService) {}
 
   public handleConnection(socket: Socket): void {
     socket.join('lobbies');
   }
 
   public handleDisconnect(socket: ExtendedSocket): void {
+    this.leaveLobby(socket);
+  }
+
+  @SubscribeMessage('leave-lobby')
+  public leaveLobby(@ConnectedSocket() socket: ExtendedSocket): void {
     const { lobbyId } = socket;
     const lobby = this.lobbyService.getLobby(lobbyId);
     if (!lobby) {
       return;
     }
 
+    const isHost = lobby.host.socketId === socket.id;
     const { host, player } = lobby.removePlayer(socket.id);
     if (player) {
       this.server.to(lobbyId).emit('player-left', player.username);
+      socket.lobbyId = null;
     }
 
-    if (host) {
+    if (isHost && host) {
       this.server.to(lobbyId).emit('host-changed', host);
-    } else {
+    } else if (isHost && !host) {
       this.lobbyService.removeLobby(lobby);
       this.server.to('lobbies').emit('lobby-removed', lobbyId);
     }
@@ -54,6 +63,7 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
     socket.lobbyId = lobby.id;
     socket.leave('lobbies');
+    socket.join(lobby.id);
 
     this.lobbyService.addLobby(lobby);
     socket.emit('create-lobby-response', lobby.id);
@@ -61,7 +71,7 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect {
   }
 
   @SubscribeMessage('join-lobby')
-  public joinLobby(@ConnectedSocket() socket: ExtendedSocket, @MessageBody() { lobbyId }: JoinLobbyPayload): void {
+  public joinLobby(@ConnectedSocket() socket: ExtendedSocket, @MessageBody() lobbyId: string): void {
     const { username } = socket.handshake.query;
     const player = new Player(username, socket.id);
 
@@ -88,8 +98,29 @@ export class LobbyGateway implements OnGatewayConnection, OnGatewayDisconnect {
   }
 
   @SubscribeMessage('get-lobby')
-  public getLobby(@ConnectedSocket() socket: Socket, @MessageBody() { lobbyId }: GetLobbyPayload): void {
+  public getLobby(@ConnectedSocket() socket: Socket, @MessageBody() lobbyId: string): void {
     const lobby = this.lobbyService.getLobby(lobbyId);
     socket.emit('get-lobby-response', lobby.toSummary());
+  }
+
+  @SubscribeMessage('start-lobby')
+  public startLobby(@ConnectedSocket() socket: ExtendedSocket): void {
+    const { lobbyId } = socket;
+    const lobby = this.lobbyService.getLobby(lobbyId);
+    if (!lobby) {
+      return;
+    }
+
+    const isHost = lobby.host.socketId === socket.id;
+    const canStart = OhHell.canStart(lobby);
+    if (!isHost || !canStart) {
+      return;
+    }
+
+    this.lobbyService.removeLobby(lobby);
+    const ohHell = new OhHell(lobby);
+    this.gameService.addGame(ohHell);
+
+    this.server.to(lobbyId).emit('lobby-starting', GameType.OH_HELL);
   }
 }
